// app/types.d.ts
export type Package = {
    "ID": string,
    "destino": string,
    "subtitulo": string,
    "date": string,
    "date2": string,
    "date3" : string,
    "dia"?: string,
    "dia2"?: string,
    "dia3"?: string,
    "mes"?: string,
    "mes2"?: string,
    "mes3"?: string,
    "year"?: string,
    "year2"?: string,
    "year3"?: string,
    "days": string,
    "nights": string,
    "regimen": string,
    "transporte": string,
    "periodo": string,
    "paquete": string,
    "descripcion": string,
    "moneda": string,
    "precio": string,
    "adicional": string,
    "image": string,
    "desde": number,
    "hotel": string,
    "incluye": string,
    "incluye2": string,
    "incluye3": string,
    "incluye4": string,
    "observaciones": string,
    "observaciones2": string,
    "observaciones3": string,
    "itinerario": string,
    "itinerario2": string,
    "itinerario3": string,
    "itinerario4": string,
    "itinerario5": string,
    "itinerario6": string,
    "itinerario7": string,
    "itinerario8": string,
    "tarifas": string,
    "tarifas2": string,
    "tarifas3": string,
    "tarifas4": string,
    "tarifas5": string
  }
export type Flyers = {
    "ID": string,
    image: string,
    descripcion: string,
    periodo: string
}